Move helpers and consts
